+++ previous-versions/commerce/RELEASES.md
@@ -659,3 +659,13 @@ Based on:
 - [typescript v0.38.0] previous-versions/commerce
 ### Releases
 - [NPM v0.38.0] https://www.npmjs.com/package/@codat/commerce/v/0.38.0 - previous-versions/commerce
+
+## 2023-10-01 00:11:02
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.92.2 (2.142.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.39.0] previous-versions/commerce
+### Releases
+- [NPM v0.39.0] https://www.npmjs.com/package/@codat/commerce/v/0.39.0 - previous-versions/commerce